Bug #1044651 fixed, NonMultipartComposer dropper, due to different interfaces of email.Message.Message childs. Now there is only the one Composer, messageClass can be any class supporting interface defined by email.Message.Message.
If body submitted to composer is an unicode string - it's encoded with message charset, we dont encode headers because all they are wrapped with clever email.Header.Header class.

qMail still doesn't support a lot of email module's features (write a different code for multipart messages with files attached), but it defines simple possibility to send plain messages with different transports.